user logout and restrictions added

+++ base/views.py
@@ -1,8 +1,12 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from .models import Room, Message, Topic
+from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
 from .forms import RoomForm, MessageForm
 from django.db.models import Q
+from django.contrib import messages
 
 # Create your views here.
 
@@ -39,7 +43,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
-
+@login_required(login_url='login')
 def CreateRoom(request):
     form = RoomForm()
 
@@ -53,11 +57,14 @@ def CreateRoom(request):
     return render(request, 'base/room_form.html',context)
 
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id= pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to do this!")
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -68,9 +75,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html',context)
 
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not authorized to do this!")
+
     if request.method =='POST':
         room.delete()
         return redirect('home')
@@ -78,9 +89,36 @@ def deleteRoom(request, pk):
     context = {'obj':room}
     return render(request, 'base/delete_room.html',context)
 
+
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except User.DoesNotExist:
+            messages.error(request, "User doesn't exist!")
+            return render(request, 'base/login_registration.html')
+
+        user = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "Invalid password")
+
     context = {}
     return render(request, 'base/login_registration.html', context)
+
+
+def logoutPage(request):
+    logout(request)
+    return redirect('home')
 
 
 # my added
